Quotidian -Wsign-conversion fixes
Signedness fixes for nvflags stuff will not be apart of this project.
I already did most of the work for that in the expand-nvflags branch,
which is currently stalled because (primarily for stability) some of the
changes in thickfold-size_t must be done prior to that becoming truly
expedient or feasible (as regards stability).