Further transition away from misguided ssize_t to ptrdiff_t
Per the POSIX standard:
> The type ssize_t shall be capable of storing values at least
> in the range [-1, {SSIZE_MAX}].

My changes were/still are in flagrant violation of this
specification, since ssize_t is *not* guaranteed to allow
sizes lower than -1. As such, I'm moving this branch over
to ptrdiff_t, which is a C89 type that actually has such
a guarantee (making it effectively a safer signed version
of size_t, and better for portability).

Upstreaming is still a LONG ways away...