size_t/ptrdiff_t transition part 13: the rest of ksh93
This is the thirteenth of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

This covers the rest of ksh93:
- All of the other headers.
- args.c
- arith.c
- deparse.c
- expand.c
- fault.c
- io.c
  - In this file ssize_t is preferable because many of the
    underlying SFIO and POSIX functions return values of
    that type.
  - sh_sfeval(): Removed the set but not used ep->slen variable.
- jobs.c
  - Get rid of if/else PID botch; a single casted strtoll is fine.
- main.c
- path.c
  - Use the order 'noreturn void' for exscript() to fix a
    compiler warning.
  - path_spawn(): Added an UNREACHABLE() to fix an instance of
    -Wunreachable-code-return.
- streval.c
- xec.c

Conspicuous change of note:
- Removed an (unsigned) cast for sizeof because it's already
  unsigned.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
555 => 248 (progression from part 12 => part 13)
Most of the warnings that remain are mere bitflag issues of tertiary importance.
Some, like those pertaining to getrlimit(2), are bugs in the underlying operating
system and cannot be fixed.
I'll note that I have a patch that expands nvflags to uint64_t which helps
quash some more warnings. But that patch isn't a part of the thickfold series
and is intended for submission sometime after this (hopefully) gets merged.

Fixes https://github.com/ksh93/ksh/issues/592