(thickfold-only): Reinstate HIGHBITI to fix mktemp failure
When the /tmp directory gets filled up with many, many temporary
directories, the mktemp builtin could fail with an error and do a basic
ERROR_exit(). This was caused by my changing of HIGHBITI and the
associated code in sfvprintf to HIGHBITS, which is evidently incorrect
here (shortening the result to 32-bit is what we want; 64-bit type
handling is already done above). Revert that change to fix the mktemp
builtin failures.

I encountered this regression during the "dogfooding" stage of
thickfold testing (cf. the crescive branch). I attempted to make a
regression test, but as of now it fails intermittently because
of deficiencies in pathtemp(). In a separate branch I'm developing
a partial backport/refactor of ksh93v- pathtemp, which will hopefully
fix that issue too.

Test (TODO: Add when intermittent failing bug is fixed):
     got=$(
             set +x
             builtin mktemp rm 2>/dev/null || exit 0
             redirect 2>&1
             masterdir=${ mktemp -d ;} || exit 1
             integer i
             for ((i=0; i!=50000; i++)) do
                     file=${ mktemp -p "$masterdir" ;} || exit 1
                     folder=${ mktemp -p "$masterdir" -d ;} || exit 1
             done
             exit 0
     )
     [[ -z $got ]] || err_exit "mktemp is unable to create 50,000 files and 50,000 directories (got $(printf %.200q "$got"))"